moved chicago room extrusion example to extrusions package
